refactor: #227 return notification component and function instead of error message
