+++ mw-webapp/src/App.tsx
@@ -1,24 +1,21 @@
+import {useEffect} from "react";
 import {Header} from "src/component/header/Header";
-import {Notification, NotificationType} from "src/component/notification/Notification";
-import {useErrorHandler} from "src/hooks/useErrorHandler";
+import {useErrorNotification} from "src/hooks/useErrorNotification";
 import {Router} from "src/router/Router";
 
 /**
  * App
  */
 export const App = () => {
-  const errorMessage = useErrorHandler();
+  const {ErrorNotification, triggerErrorNotification} = useErrorNotification();
+
+  useEffect(() => {
+    triggerErrorNotification("test error", Date.now());
+  }, []);
 
   return (
     <>
-      {errorMessage &&
-        <Notification
-          content={<div>
-            {errorMessage}
-          </div>}
-          type={NotificationType.foreground}
-        />
-      }
+      {ErrorNotification}
       <Header />
       <Router />
     </>
